+++ openadapt/models.py
@@ -7,15 +7,17 @@ from typing import Any, Type, Union
 import copy
 import io
 import sys
+import textwrap
 
 from bs4 import BeautifulSoup
-from oa_pynput import keyboard
+from pynput import keyboard
 from PIL import Image, ImageChops
 import numpy as np
 import sqlalchemy as sa
 
 from openadapt.config import config
 from openadapt.custom_logger import logger
+from openadapt.drivers import anthropic
 from openadapt.db import db
 from openadapt.privacy.base import ScrubbingProvider, TextScrubbingMixin
 from openadapt.privacy.providers import ScrubProvider
@@ -110,6 +112,9 @@ class Recording(db.Base):
         if not self._processed_action_events:
             session = crud.get_new_session(read_only=True)
             self._processed_action_events = events.get_events(session, self)
+            # Preload screenshots to avoid lazy loading later
+            for event in self._processed_action_events:
+                event.screenshot
         return self._processed_action_events
 
     def scrub(self, scrubber: ScrubbingProvider) -> None:
@@ -125,6 +130,7 @@ class ActionEvent(db.Base):
     """Class representing an action event in the database."""
 
     __tablename__ = "action_event"
+    _repr_ignore_attrs = ["reducer_names"]
 
     _segment_description_separator = ";"
 
@@ -333,6 +339,11 @@ class ActionEvent(db.Base):
         if not value == self.canonical_text:
             logger.warning(f"{value=} did not match {self.canonical_text=}")
 
+    @property
+    def raw_text(self) -> str:
+        """Return a string containing the raw action text (without separators)."""
+        return "".join(self.text.split(config.ACTION_TEXT_SEP))
+
     def __str__(self) -> str:
         """Return a string representation of the action event."""
         attr_names = [
@@ -492,10 +503,6 @@ class ActionEvent(db.Base):
         Returns:
             dictionary containing relevant properties from the ActionEvent.
         """
-        if self.active_browser_element:
-            import ipdb
-
-            ipdb.set_trace()
         action_dict = deepcopy(
             {
                 key: val
@@ -518,6 +525,9 @@ class ActionEvent(db.Base):
             action_dict["available_segment_descriptions"] = (
                 self.available_segment_descriptions
             )
+
+        # TODO:
+        """
         if self.active_browser_element:
             action_dict["active_browser_element"] = str(self.active_browser_element)
         if self.available_browser_elements:
@@ -526,11 +536,7 @@ class ActionEvent(db.Base):
             action_dict["available_browser_elements"] = str(
                 self.available_browser_elements
             )
-
-        if self.active_browser_element:
-            import ipdb
-
-            ipdb.set_trace()
+        """
         return action_dict
 
     @property
@@ -548,6 +554,76 @@ class ActionEvent(db.Base):
             return self.recording.action_events[current_index + 1]
 
         return None
+
+    def prompt_for_description(self, return_image: bool = False) -> str:
+        """Use the Anthropic API to describe what is happening in the action event.
+
+        Args:
+            return_image (bool): Whether to return the image sent to the model.
+
+        Returns:
+            str: The description of the action event.
+        """
+        from openadapt.plotting import display_event
+
+        image = display_event(
+            self,
+            marker_width_pct=0.05,
+            marker_height_pct=0.05,
+            darken_outside=0.7,
+            display_text=False,
+            marker_fill_transparency=0,
+            dim_outside_window=False,
+        )
+
+        if self.text:
+            description = f"Type '{self.raw_text}'"
+        else:
+            prompt = (
+                "What user interface element is contained in the highlighted circle "
+                "of the image?"
+            )
+            # TODO: disambiguate
+            system_prompt = textwrap.dedent(
+                """
+                Briefly describe the user interface element in the screenshot at the
+                highlighted location.
+                For example:
+                - "OK button"
+                - "URL bar"
+                - "Down arrow"
+                DO NOT DESCRIBE ANYTHING OUTSIDE THE HIGHLIGHTED AREA.
+                Do not append anything like "is contained within the highlighted circle
+                in the calculator interface." Just name the user interface element.
+            """
+            )
+
+            logger.info(f"system_prompt=\n{system_prompt}")
+            logger.info(f"prompt=\n{prompt}")
+
+            # Call the Anthropic API
+            element = anthropic.prompt(
+                prompt=prompt,
+                system_prompt=system_prompt,
+                images=[image],
+            )
+
+            if self.name == "move":
+                description = f"Move mouse to '{element}'"
+            elif self.name == "scroll":
+                # TODO: "scroll to", dx/dy
+                description = f"Scroll mouse on '{element}'"
+            elif "click" in self.name:
+                description = (
+                    f"{self.mouse_button_name.capitalize()} {self.name} '{element}'"
+                )
+            else:
+                raise ValueError(f"Unhandled {self.name=} {self}")
+
+        if return_image:
+            return description, image
+        else:
+            return description
 
 
 class WindowEvent(db.Base):
@@ -654,9 +730,9 @@ class WindowEvent(db.Base):
                     "title",
                     "help",
                 ]
-                if sys.platform == "win32":
+                if sys.platform != "darwin":
                     logger.warning(
-                        "key_suffixes have not yet been defined on Windows."
+                        "key_suffixes have not yet been defined on {sys.platform=}."
                         "You can help by uncommenting the lines below and pasting "
                         "the contents of the window_dict into a new GitHub Issue."
                     )